llm_commercial_project/2025.9.25 enterprise-rag-system/requirements.txt
# 基础依赖
python==3.12
torch==2.7.0+cu128
transformers==4.41.2
sentence-transformers==3.0.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-multipart==0.0.9

# 向量数据库
milvus==2.3.4
pymilvus==2.3.5

# 文档处理
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==0.6.23
markdown==3.6
nltk==3.8.1
langchain==0.2.5
langchain-community==0.2.5

# 高级检索
rank-bm25==0.2.2
tiktoken==0.7.0

# 监控与日志
prometheus-client==0.20.0
python-json-logger==2.0.7

# 部署
docker==6.1.3
docker-compose==1.29.2
